refactor: move expand_rv64_limbs to shared riscv-circuit adapters
Move the helper from keccak256 lib.rs to openvm_riscv_circuit::adapters
so it can be reused by deferral, sha2, and other extensions that need
to expand word-size limbs to register-size with zero padding.
