+++ extensions/keccak256/circuit/src/lib.rs
@@ -21,19 +21,5 @@ mod constants;
 mod extension;
 pub use constants::*;
 pub use extension::*;
-use openvm_instructions::riscv::{RV64_REGISTER_NUM_LIMBS, RV64_WORD_NUM_LIMBS};
-use openvm_stark_backend::p3_field::PrimeCharacteristicRing;
-
-/// Expand RV64_WORD_NUM_LIMBS (4) limbs to RV64_REGISTER_NUM_LIMBS (8) by zero-padding the upper
-/// limbs. Used for register bus reads where the register holds a 32-bit value in the low 4 bytes.
-pub fn expand_rv64_limbs<V: Copy + Into<T>, T: PrimeCharacteristicRing>(
-    limbs: &[V; RV64_WORD_NUM_LIMBS],
-) -> [T; RV64_REGISTER_NUM_LIMBS] {
-    std::array::from_fn(|i| {
-        if i < RV64_WORD_NUM_LIMBS {
-            limbs[i].into()
-        } else {
-            T::ZERO
-        }
-    })
-}
+// Re-export for use by keccakf_op and xorin AIRs.
+pub use openvm_riscv_circuit::adapters::expand_rv64_limbs;
